requirements.txt
PyYAML
pandas
fastparquet